Cleanup HTTP URLs to use HTTPS.
This commit updates URLs to prefer the https protocol. Redirects are not followed to avoid accidentally expanding intentionally shortened URLs (i.e. if using a URL shortener).

These URLs were fixed, but the https status was not OK. However, the https status was the same as the http request or http redirected to an https URL, so they were migrated. Your review is recommended.

* http://docs.spring.io/spring-boot-data-geode/docs/ (301) migrated to:
  https://docs.spring.io/spring-boot-data-geode/docs/ ([https](https://docs.spring.io/spring-boot-data-geode/docs/) result 404).

These URLs were fixed successfully.

* http://docs.spring.io/spring-data/commons/docs/current/api/ migrated to:
  https://docs.spring.io/spring-data/commons/docs/current/api/ ([https](https://docs.spring.io/spring-data/commons/docs/current/api/) result 200).
* http://docs.spring.io/spring/docs/current/javadoc-api/ migrated to:
  https://docs.spring.io/spring/docs/current/javadoc-api/ ([https](https://docs.spring.io/spring/docs/current/javadoc-api/) result 200).
* http://geode.apache.org/releases/latest/javadoc/ migrated to:
  https://geode.apache.org/releases/latest/javadoc/ ([https](https://geode.apache.org/releases/latest/javadoc/) result 200).

Resolves gh-25.

Resolves gh-27.