Ensure Asciidoctor attributes are configured for PDFs too.
In previous versions of the Asciidoctor Gradle Plugin, Asciidoctor was an extension that was read by all Asciidoctor
tasks. This meant that configuring the following impacted all Asciidoctor tasks (including PDF):

asciidoctor {
   attributes ...
}

The latest version of Asciidoctor moved conventions to AsciidoctorJ. The reason for this is there was a task named
'asciidoctor' and a convention named 'asciidoctor' which caused confusion for users on which was being configured.

This commit switches to the new 'asciidoctorj' convention/extension so that attributes are configured for all
Asciidoctor tasks (i.e. both the 'asciidoctor' and 'asciidoctorPdf' tasks).