Refactor the Async Inline Caching Spring Geode Sample Spring Boot application (code).
Change the default AEQ batch size to 25 and enable AEQ batch conflation.

Synchronize the Set of players (Golfers) and List of Pairings in the GolfTournament class.

Add a getPlayers() method to the GolfTournament class returning the Golfers registered to play in the GolfTournament.

Simply the isFinished() method in the GolfTournament class.

Add the in(:Golfer) and signScorecard() methods to the GolfTournament.Pairing class.

Rename the GolfTournamentService class to PgaTourService.

Rename the GolfTournamentController class to GolferController.