Edit and complete the Async Inline Caching Sample Guide.
Refactor the Async Inline Caching Sample Code to include (Asciidoc) documentation markers.

Change the batch time interval in the AEQ batch-size configuration/strategy bean definition to 15 minutes.

Rename the GolfTournament created in the SpringBootApplication class to 'The Masters'.

Refactor the isFinished(..) and isNotFinished() methods in the PgaTourService class to accept a nullable GolfTournament argument.

Refactor the GolferController class to declare the REST API web service endpoints to use '/api/golf/tournament/' as the base Webapp context.