Refactor the @EnableClusterAware configuration and conditional logic to determine whether an Apache Geode based cluster is available or not.
Applied Spring's @NonNull and @Nullable annotations to the API.

Added a NotKubernetesEnvironmentCondition class to the ClusterNotAvailableConfiguration class.

Added a @Conditional on the NotKubernetesEnvironmentCondition class annotated IsNotKubernetesEnvironmentCondition class to the ClusterNotAvailableConfiguration.AllClusterNotAvailableConditions class.

Removed all @ConditionalOnCloudPlatform annotated Condition classes in the ClusterAvailableConfiguration.AnyClusterAvailableCondition class, for both CloudFoundry and Kubernetes.

Edit Javadoc.

Resolves gh-99.